Optimization Algorithm for MRTA
Here input is taken as number of robots and task which is equla to N.
The random generator generates distances between each robot and tasks in
a matrix which is located in space of (0,0) to (500,500).
The hungarian, robot and stable-marriage genrate matching between them
and Normalize generates a normalized matrix of original on which Robot
algorithm can work on.
The master has the task to run all these files and generate results for
N ranging from 10 to 100 and each N has 25 random matrix to work on.